Moved tagging demos to the lab section until they're considered stable (Resolves #1371)
