+++ super_editor/example/lib/demos/in_the_lab/feature_stable_tags.dart
@@ -1,8 +1,9 @@
-import 'package:example/demos/features/feature_demo_scaffold.dart';
-import 'package:example/demos/features/popover_list.dart';
+import 'package:example/demos/in_the_lab/in_the_lab_scaffold.dart';
 import 'package:flutter/material.dart' hide ListenableBuilder;
 import 'package:follow_the_leader/follow_the_leader.dart';
 import 'package:super_editor/super_editor.dart';
+
+import 'popover_list.dart';
 
 class UserTagsFeatureDemo extends StatefulWidget {
   const UserTagsFeatureDemo({super.key});
@@ -94,17 +95,21 @@ class _UserTagsFeatureDemoState extends State<UserTagsFeatureDemo> {
   Widget build(BuildContext context) {
     return Stack(
       children: [
-        FeatureDemoScaffold(
-          child: Row(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: [
-              Expanded(
-                child: _buildEditor(),
-              ),
-              _buildTagList(),
-            ],
-          ),
+        InTheLabScaffold(
+          content: _buildEditor(),
+          supplemental: _buildTagList(),
         ),
+        // FeatureDemoScaffold(
+        //   child: Row(
+        //     crossAxisAlignment: CrossAxisAlignment.stretch,
+        //     children: [
+        //       Expanded(
+        //         child: _buildEditor(),
+        //       ),
+        //       _buildTagList(),
+        //     ],
+        //   ),
+        // ),
         if (_userTagPlugin.tagIndex.composingStableTag.value != null)
           Follower.withOffset(
             link: _composingLink,
@@ -123,87 +128,75 @@ class _UserTagsFeatureDemoState extends State<UserTagsFeatureDemo> {
   }
 
   Widget _buildEditor() {
-    return SuperEditor(
-      editor: _editor,
-      document: _document,
-      composer: _composer,
-      focusNode: _editorFocusNode,
-      stylesheet: defaultStylesheet.copyWith(
-        inlineTextStyler: (attributions, existingStyle) {
-          TextStyle style = defaultInlineTextStyler(attributions, existingStyle);
+    return IntrinsicHeight(
+      child: SuperEditor(
+        editor: _editor,
+        document: _document,
+        composer: _composer,
+        focusNode: _editorFocusNode,
+        stylesheet: defaultStylesheet.copyWith(
+          inlineTextStyler: (attributions, existingStyle) {
+            TextStyle style = defaultInlineTextStyler(attributions, existingStyle);
 
-          if (attributions.contains(stableTagComposingAttribution)) {
-            style = style.copyWith(
-              color: Colors.blue,
-            );
-          }
+            if (attributions.contains(stableTagComposingAttribution)) {
+              style = style.copyWith(
+                color: Colors.blue,
+              );
+            }
 
-          if (attributions.whereType<CommittedStableTagAttribution>().isNotEmpty) {
-            style = style.copyWith(
-              color: Colors.orange,
-            );
-          }
+            if (attributions.whereType<CommittedStableTagAttribution>().isNotEmpty) {
+              style = style.copyWith(
+                color: Colors.orange,
+              );
+            }
 
-          return style;
-        },
-        addRulesAfter: [
-          ...darkModeStyles,
-        ],
-      ),
-      documentOverlayBuilders: [
-        AttributedTextBoundsOverlay(
-          selector: (a) => a == stableTagComposingAttribution,
-          builder: (context, attribution) {
-            return Leader(
-              link: _composingLink,
-              child: DecoratedBox(
-                decoration: BoxDecoration(
-                  border: Border.all(color: Colors.blue),
-                ),
-              ),
-            );
+            return style;
           },
+          addRulesAfter: [
+            ...darkModeStyles,
+          ],
         ),
-        DefaultCaretOverlayBuilder(
-          caretStyle: CaretStyle().copyWith(color: Colors.redAccent),
-        ),
-      ],
-      plugins: {
-        _userTagPlugin,
-      },
+        documentOverlayBuilders: [
+          AttributedTextBoundsOverlay(
+            selector: (a) => a == stableTagComposingAttribution,
+            builder: (context, attribution) {
+              return Leader(
+                link: _composingLink,
+                child: DecoratedBox(
+                  decoration: BoxDecoration(
+                    border: Border.all(color: Colors.blue),
+                  ),
+                ),
+              );
+            },
+          ),
+          DefaultCaretOverlayBuilder(
+            caretStyle: CaretStyle().copyWith(color: Colors.redAccent),
+          ),
+        ],
+        plugins: {
+          _userTagPlugin,
+        },
+      ),
     );
   }
 
   Widget _buildTagList() {
-    return Container(
-      width: 300,
-      decoration: BoxDecoration(
-        border: Border(
-          left: BorderSide(width: 1, color: Colors.white.withOpacity(0.1)),
+    if (_users.isEmpty) {
+      return const SizedBox();
+    }
+
+    return Center(
+      child: SingleChildScrollView(
+        child: Wrap(
+          spacing: 12,
+          runSpacing: 12,
+          alignment: WrapAlignment.center,
+          children: [
+            for (final tag in _users) //
+              Chip(label: Text(tag)),
+          ],
         ),
-      ),
-      padding: const EdgeInsets.all(24),
-      child: Center(
-        child: _users.isNotEmpty
-            ? SingleChildScrollView(
-                child: Wrap(
-                  spacing: 12,
-                  runSpacing: 12,
-                  alignment: WrapAlignment.center,
-                  children: [
-                    for (final tag in _users) //
-                      Chip(label: Text(tag)),
-                  ],
-                ),
-              )
-            : Text(
-                "NO USERS",
-                style: TextStyle(
-                  color: Colors.white.withOpacity(0.1),
-                  fontSize: 32,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
       ),
     );
   }
